temp: visualise
Implement running over all Semver lints. There is one issue left in the
visualisation algorithm regarding YieldFrom followed by
AdvanceInputIterator, and then all that remains is to deal with overhead
and leftover advance_iterators.

With any luck, I should have visualisation fully functional by tomorrow.

From here, the remaining work consists of designing and integrating the
data model. For example, do we want information about YieldInto? I know
Predrag has mentioned wanting a breakdown of a query's components, is
that something we should still look into? Also, there is more abstract
information at the graph level, such as edge traversals. I'm not sure
how this information could be used, but I do know Predrag mentioned
wanting it, so I'll have to follow up on that. It would be lovely to
test the visualisation against trustfall-rustdoc-adapter#676 so we
can see which lints fail to benefit and why and hopefully get that change
merged.

Additionally, I want to spend some portion of time after the algorithm
is complete working back through all these notes and summarising the
experience for reference during the review portions of GSoC and
potentially as material for a blog post.
Once this is done, the code needs to be cleaned up into a workable form
for sharing. At this stage, I want to leave visualise.py as a python
file because it's easy to work with and fast to iterate from, but
down the track it will need to become a rust script.